refactor: split mutationHandlers, structureHandlers, and useFileOperations (partial #94)
- mutationHandlers.ts (774 lines) -> batchEditHandler.ts, applyDiffHandler.ts,
  replaceAnchoredHandler.ts with barrel re-export
- structureHandlers.ts (750 lines) -> astHandlers.ts (shared types/utils + AST/digest)
  and blockHandlers.ts (list blocks, resolve targets, get section)
- useFileOperations.ts (701 lines) -> useFileSave.ts, useFileOpen.ts,
  useFileShortcuts.ts with thin orchestrator hook
