+++ src/hooks/useFileShortcuts.ts
@@ -1,0 +1,143 @@
+/**
+ * File Shortcuts — Menu Event Listeners and Keyboard Shortcuts
+ *
+ * Purpose: Wire up Tauri menu events and keyboard shortcuts for file operations.
+ *   Registers listeners for menu:new, menu:open, menu:save, menu:save-as,
+ *   menu:move-to, menu:save-all-quit, and open-file events. Also handles
+ *   direct keyboard shortcuts for Save and Save As.
+ *
+ * @coordinates-with useFileSave.ts — save/saveAs/moveTo/saveAllQuit handlers
+ * @coordinates-with useFileOpen.ts — open/openFile/new handlers
+ * @coordinates-with useFileOperations.ts — main hook that calls this
+ * @module hooks/useFileShortcuts
+ */
+
+import { useEffect, useRef } from "react";
+import { type UnlistenFn } from "@tauri-apps/api/event";
+import { getCurrentWebviewWindow } from "@tauri-apps/api/webviewWindow";
+import { useShortcutsStore } from "@/stores/shortcutsStore";
+import { matchesShortcutEvent } from "@/utils/shortcutMatch";
+import { isImeKeyEvent } from "@/utils/imeGuard";
+import { safeUnlistenAll } from "@/utils/safeUnlisten";
+
+import { handleSave, handleSaveAs, handleMoveTo, handleSaveAllQuit } from "./useFileSave";
+import { handleOpen, handleOpenFile, handleNew } from "./useFileOpen";
+
+/**
+ * Set up all file operation menu listeners and keyboard shortcuts.
+ * Returns a cleanup function.
+ */
+export function useFileShortcuts(windowLabel: string): void {
+  const unlistenRefs = useRef<UnlistenFn[]>([]);
+
+  useEffect(() => {
+    let cancelled = false;
+
+    const setupListeners = async () => {
+      // Clean up any existing listeners first
+      unlistenRefs.current = safeUnlistenAll(unlistenRefs.current);
+
+      if (cancelled) return;
+
+      // Get current window for filtering - menu events include target window label
+      const currentWindow = getCurrentWebviewWindow();
+
+      // menu:new creates a new tab in this window
+      // menu:new-window (handled in Rust) creates a new window
+      const unlistenNew = await currentWindow.listen<string>("menu:new", (event) => {
+        if (event.payload !== windowLabel) return;
+        handleNew(windowLabel);
+      });
+      if (cancelled) { unlistenNew(); return; }
+      unlistenRefs.current.push(unlistenNew);
+
+      const unlistenOpen = await currentWindow.listen<string>("menu:open", async (event) => {
+        if (event.payload !== windowLabel) return;
+        await handleOpen(windowLabel);
+      });
+      if (cancelled) { unlistenOpen(); return; }
+      unlistenRefs.current.push(unlistenOpen);
+
+      const unlistenSave = await currentWindow.listen<string>("menu:save", async (event) => {
+        console.log("[FileOps] menu:save event received, payload:", event.payload);
+        if (event.payload !== windowLabel) return;
+        await handleSave(windowLabel);
+      });
+      if (cancelled) { unlistenSave(); return; }
+      unlistenRefs.current.push(unlistenSave);
+
+      const unlistenSaveAs = await currentWindow.listen<string>("menu:save-as", async (event) => {
+        if (event.payload !== windowLabel) return;
+        await handleSaveAs(windowLabel);
+      });
+      if (cancelled) { unlistenSaveAs(); return; }
+      unlistenRefs.current.push(unlistenSaveAs);
+
+      const unlistenMoveTo = await currentWindow.listen<string>("menu:move-to", async (event) => {
+        if (event.payload !== windowLabel) return;
+        await handleMoveTo(windowLabel);
+      });
+      if (cancelled) { unlistenMoveTo(); return; }
+      unlistenRefs.current.push(unlistenMoveTo);
+
+      // Save All and Quit - saves all dirty documents then quits
+      const unlistenSaveAllQuit = await currentWindow.listen<string>(
+        "menu:save-all-quit",
+        async (event) => {
+          if (event.payload !== windowLabel) return;
+          await handleSaveAllQuit(windowLabel);
+        }
+      );
+      if (cancelled) { unlistenSaveAllQuit(); return; }
+      unlistenRefs.current.push(unlistenSaveAllQuit);
+
+      // Listen for open-file from FileExplorer (window-local event, payload contains path)
+      const unlistenOpenFile = await currentWindow.listen<{ path: string }>(
+        "open-file",
+        async (event) => {
+          await handleOpenFile(windowLabel, event.payload.path);
+        }
+      );
+      if (cancelled) { unlistenOpenFile(); return; }
+      unlistenRefs.current.push(unlistenOpenFile);
+    };
+
+    setupListeners();
+
+    // Keyboard shortcut handler for file operations
+    // Menu accelerators don't always work reliably (TipTap captures events),
+    // so we listen directly for Save and Save As shortcuts.
+    const handleKeyDown = (e: KeyboardEvent) => {
+      if (isImeKeyEvent(e)) return;
+      const target = e.target as HTMLElement;
+      if (target.tagName === "INPUT" || target.tagName === "TEXTAREA") return;
+
+      const shortcuts = useShortcutsStore.getState();
+
+      // Save As (Cmd+Shift+S)
+      const saveAsKey = shortcuts.getShortcut("saveAs");
+      if (matchesShortcutEvent(e, saveAsKey)) {
+        e.preventDefault();
+        handleSaveAs(windowLabel);
+        return;
+      }
+
+      // Save (Cmd+S)
+      const saveKey = shortcuts.getShortcut("save");
+      if (matchesShortcutEvent(e, saveKey)) {
+        console.log("[FileOps] Cmd+S keyboard shortcut matched");
+        e.preventDefault();
+        handleSave(windowLabel);
+        return;
+      }
+    };
+
+    window.addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      cancelled = true;
+      unlistenRefs.current = safeUnlistenAll(unlistenRefs.current);
+      window.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [windowLabel]);
+}
